Moved simulations to their own files
