+++ internal/core/animations.rs
@@ -9,7 +9,7 @@ use core::cell::Cell;
 #[cfg(not(feature = "std"))]
 use num_traits::Float;
 
-pub(crate) mod physics_simulation;
+pub(crate) mod simulations;
 
 mod cubic_bezier {
     //! This is a copy from lyon_algorithms::geom::cubic_bezier implementation
@@ -323,10 +323,10 @@ fn ease_out_bounce_curve(value: f32) -> f32 {
 
 /// Evaluates a mass/stiffness/damping spring at `elapsed_secs`, returning `(progress, settled)`.
 pub fn spring_settle_progress(
-    regime: &physics_simulation::SpringRegime,
+    regime: &simulations::spring::SpringRegime,
     elapsed_secs: f32,
 ) -> (f32, bool) {
-    use physics_simulation::{SPRING_SETTLE_POSITION_EPSILON, SPRING_SETTLE_VELOCITY_EPSILON};
+    use simulations::spring::{SPRING_SETTLE_POSITION_EPSILON, SPRING_SETTLE_VELOCITY_EPSILON};
 
     let (rel_pos, rel_vel) = regime.evaluate(elapsed_secs);
     let settled = rel_pos.abs() < SPRING_SETTLE_POSITION_EPSILON
